requirements.txt
python-dotenv==0.15.0
aniso8601==8.1.0
astroid==2.4.2
blinker==1.4
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
flake8==3.8.4
Flask==1.1.2
Flask-HTTPAuth==4.2.0
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.7.3
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.4.2
pylint==2.6.0
pyperclip==1.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.25.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.2
typed-ast==1.4.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.0
